feat!(svg): move svg structures to private header
BREAKING CHANGE: this breaks user code that might be relying on
the internal SVG structures. The migration path for these users
is to use the private API by either including "lvgl_private.h" or
enabling LV_USE_PRIVATE_API
