+++ src/lvgl_private.h
@@ -142,6 +142,7 @@
 #include "libs/rle/lv_rle_private.h"
 #include "libs/rlottie/lv_rlottie_private.h"
 #include "image/svg/lv_svg_decoder.h"
+#include "image/svg/lv_svg_private.h"
 #include "image/svg/lv_svg_parser.h"
 #include "image/svg/lv_svg_render.h"
 #include "image/svg/lv_svg_token.h"
